Renamed files and moved into different folders to follow standard
