+++ src/router.dispatch.ts
@@ -1,11 +1,12 @@
 // Do Router dispatch here, i.e. map incoming routes to appropriate router
 import { Application, Request, Response } from "express";
 import indexRouter from "./routers/indexRouter";
-import soleTraderRouter from "./routers/soleTraderRouter";
+import dateOfBirthRouter from "./routers/soleTraderDateOfBirthRouter";
 
 const routerDispatch = (app: Application) => {
+    const soleTraderRoute : string = "/sole-trader";
     app.use("/", indexRouter);
-    app.use("/sole-trader", soleTraderRouter);
+    app.use(soleTraderRoute, dateOfBirthRouter);
     app.use("*", (req: Request, res: Response) => {
         res.status(404).render("partials/error_400");
     });
